requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.4.3
langchain==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5
langchain-core==0.2.9
openai==1.35.3
